refactor: rename select to subset.py in process for ruff check
